requirements.txt

streamlit
python-dotenv
langchain
openai
PyPDF2

